btrfs: Avoid unnecessary loops when initializing data
Previous commit 9037c6d608ce fixed the problem that the
btrfs devices can be hidden in ssm list if the device itself was the
type we hide by default.
However it's better to do it when we discover the device, rather then
loop over the devices again once we're done.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Fixes: 9037c6d ("btrfs: unhide devices in tests")